requirements.txt
python-telegram-bot==13.3
peewee==3.14.1
validators==0.18.2